requirements.txt
adal==1.2.7
argcomplete==1.12.3
azure-common==1.1.27
azure-core==1.20.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==19.0.0
azure-mgmt-storage==19.0.0
azure-ml==0.0.1
azureml-core==1.37.0.post1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
contextlib2==21.6.0
cryptography==3.4.8
docker==5.0.3
humanfriendly==9.2
idna==3.3
isodate==0.6.1
jeepney==0.7.1
jmespath==0.10.0
jsonpickle==2.0.0
knack==0.8.2
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
oauthlib==3.1.1
packaging==21.3
paramiko==2.8.1
pathspec==0.9.0
pkginfo==1.8.2
portalocker==2.3.2
pyasn1==0.4.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
SecretStorage==3.3.1
six==1.16.0
tabulate==0.8.9
urllib3==1.26.7
websocket-client==1.2.3
